Implement 'auth' route
I've separated the authentication process out into a separate blueprint. This should allow me to force users trying to access any route to authenticate first.

In addition, I could change the authentication section and not have to go fiddling around with everything else in the application
