+++ app/auth/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+auth_blueprint = Blueprint('auth', __name__)
+
+from app.auth import routes
